+++ resharper/resharper-unity/src/Settings/NamespaceProviderProjectSettingsProvider.cs
@@ -30,32 +30,64 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
             ExcludeFolderFromNamespace(mountPoint, "Assets");
             ExcludeFolderFromNamespace(mountPoint, @"Assets\Scripts");
 
-            // Don't require Packages, or the immediate children of Packages in namespaces. The immediate children are
-            // package names, and shouldn't be considered part of the namespace.
-            // E.g. Given Packages/com.foo.package/package.json don't require Packages.com.foo.packages in the namespace
-            ExcludeFolderFromNamespace(mountPoint, "Packages");
-            ExcludeImmediateSubfoldersFromNamespace(mountPoint, project, "Packages");
-
-            ExcludeFolderFromNamespace(mountPoint, "Library");
-            ExcludeFolderFromNamespace(mountPoint, @"Library\PackageCache");
+            ExcludePackagesFoldersFromNamespace(mountPoint, project, "Packages");
             foreach (var projectFolder in project.GetSubFolders("Library"))
-                ExcludeImmediateSubfoldersFromNamespace(mountPoint, projectFolder, "PackageCache", @"Library\");
+            {
+                ExcludeFolderFromNamespace(mountPoint, "Library");
+                ExcludePackagesFoldersFromNamespace(mountPoint, projectFolder, "PackageCache", @"Library");
+            }
         }
 
         private void ExcludeFolderFromNamespace(ISettingsStorageMountPoint mountPoint, string path)
         {
-            var index = NamespaceFolderProvider.GetIndexFromOldIndex(FileSystemPath.Parse(path, FileSystemPathInternStrategy.TRY_GET_INTERNED_BUT_DO_NOT_INTERN));
+            var index = NamespaceFolderProvider.GetIndexFromOldIndex(FileSystemPath.Parse(path,
+                FileSystemPathInternStrategy.TRY_GET_INTERNED_BUT_DO_NOT_INTERN));
             SetIndexedValue(mountPoint, NamespaceProviderSettingsAccessor.NamespaceFoldersToSkip, index, true);
         }
 
-        private void ExcludeImmediateSubfoldersFromNamespace(ISettingsStorageMountPoint mountPoint, IProjectFolder root, string parent, string rootPath = "")
+        private void ExcludePackagesFoldersFromNamespace(ISettingsStorageMountPoint mountPoint,
+                                                         IProjectFolder packageHierarchyParentFolder,
+                                                         string packageHierarchyFolderName,
+                                                         string parentProjectRelativePath = "")
         {
-            foreach (var projectFolder in root.GetSubFolders(parent))
+            var packageHierarchyRootFolderPath = parentProjectRelativePath.Length == 0
+                ? packageHierarchyFolderName
+                : parentProjectRelativePath + @"\" + packageHierarchyFolderName;
+
+            foreach (var subFolder in packageHierarchyParentFolder.GetSubFolders(packageHierarchyFolderName))
             {
-                foreach (var subFolder in projectFolder.GetSubFolders())
+                // Exclude the root of the package hierarchy (Packages or PackageCache)
+                ExcludeFolderFromNamespace(mountPoint, packageHierarchyRootFolderPath);
+                ExcludePackageSubFoldersFromNamespace(mountPoint, subFolder, packageHierarchyRootFolderPath);
+            }
+        }
+
+        private void ExcludePackageSubFoldersFromNamespace(ISettingsStorageMountPoint mountPoint,
+                                                           IProjectFolder thisFolder, string thisPath)
+        {
+            // First time called, this excludes package name, e.g. `Packages/com.unity.mathematics` or
+            // `Library/PackageCache/com.unity.collections@0.0.9-preview.9`
+            // If that folder does not have a package.json recurse until we find one, and exclude everything along the
+            // way. This handles an edge case where e.g. a GitHub repo is checked out into Packages, but the root of the
+            // project isn't the package. Instead, the package is defined one or two levels down, and included into the
+            // project via file: in the manifest.
+            // E.g. myrepo/src/mypackage/{package.json}. We don't want `myrepo` or `src` appearing in the namespace
+            foreach (var subFolder in thisFolder.GetSubFolders())
+            {
+                var path = thisPath + @"\" + subFolder.Name;
+                ExcludeFolderFromNamespace(mountPoint, path);
+
+                // Is it a package folder? Exclude Scripts, Runtime, Scripts/Runtime and Runtime/Scripts
+                if (subFolder.Location.Combine("package.json").ExistsFile)
                 {
-                    ExcludeFolderFromNamespace(mountPoint, rootPath + projectFolder.Name + @"\" + subFolder.Name);
+                    ExcludeFolderFromNamespace(mountPoint, path + @"\Runtime");
+                    ExcludeFolderFromNamespace(mountPoint, path + @"\Scripts");
+                    ExcludeFolderFromNamespace(mountPoint, path + @"\Runtime\Scripts");
+                    ExcludeFolderFromNamespace(mountPoint, path + @"\Scripts\Runtime");
+                    return;
                 }
+
+                ExcludePackageSubFoldersFromNamespace(mountPoint, subFolder, path);
             }
         }